requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
flake8==7.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.1.0
propcache==0.2.1
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
yarl==1.18.3
